Single-Precision Everywhere (#21)
Bench: 3453580

+++ src/search/mod.rs
@@ -6,6 +6,6 @@ pub mod value;
 
 use params::TunableParams;
 
-pub fn cp_wdl(score: i32, params: &TunableParams) -> f64 {
-    1.0 / (1.0 + (-f64::from(score) / (100.0 * params.scale())).exp())
+pub fn cp_wdl(score: i32, params: &TunableParams) -> f32 {
+    1.0 / (1.0 + (-(score as f32) / (100.0 * params.scale())).exp())
 }
